// src/orders/order-status.enum.ts
export enum OrderStatus {
  PENDING_PAYMENT = 'PENDING_PAYMENT', 
  PAID = 'PAID', 
  SEARCHING = 'SEARCHING',
  ASSIGNED = 'ASSIGNED',
  IN_PROGRESS = 'IN_PROGRESS',
  WAITING_CONFIRMATION = 'WAITING_CONFIRMATION',
  PAYMENT_REQUIRED = 'PAYMENT_REQUIRED',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED',
  AUTO_CANCELLED = 'AUTO_CANCELLED',
  TIMEOUT = 'TIMEOUT',
}